refactor(ramp): remove buy and sell folders
