+++ app/components/Nav/Main/MainNavigator.js
@@ -49,15 +49,15 @@ import Drawer from '../../UI/Drawer';
 
 import { RampSDKProvider } from '../../UI/Ramp/common/sdk';
 import { RampType } from '../../UI/Ramp/common/types';
-import GetStarted from '../../UI/Ramp/buy/Views/GetStarted';
-import PaymentMethods from '../../UI/Ramp/buy/Views/PaymentMethods/PaymentMethods';
-import BuildQuote from '../../UI/Ramp/buy/Views/BuildQuote/BuildQuote';
-import Quotes from '../../UI/Ramp/buy/Views/Quotes';
-import CheckoutWebView from '../../UI/Ramp/buy/Views/Checkout';
+import GetStarted from '../../UI/Ramp/common/Views/GetStarted';
+import PaymentMethods from '../../UI/Ramp/common/Views/PaymentMethods/PaymentMethods';
+import BuildQuote from '../../UI/Ramp/common/Views/BuildQuote/BuildQuote';
+import Quotes from '../../UI/Ramp/common/Views/Quotes';
+import CheckoutWebView from '../../UI/Ramp/common/Views/Checkout';
 import RampSettings from '../../UI/Ramp/common/Views/Settings';
 import NetworkSwitcher from '../../UI/Ramp/common/Views/NetworkSwitcher';
 import RampAddActivationKey from '../../UI/Ramp/common/Views/Settings/AddActivationKey';
-import Regions from '../../UI/Ramp/buy/Views/Regions';
+import Regions from '../../UI/Ramp/common/Views/Regions';
 import { colors as importedColors } from '../../../styles/common';
 import OrderDetails from '../../UI/Ramp/common/Views/OrderDetails';
 import SendTransaction from '../../UI/Ramp/common/Views/SendTransaction';
