+++ app/components/UI/Ramp/common/Views/GetStarted/GetStarted.styles.ts
