+++ app/components/UI/Ramp/common/Views/GetStarted/GetStarted.tsx
@@ -3,15 +3,15 @@ import { Image, View, ScrollView } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
 import Text from '../../../../../Base/Text';
 import StyledButton from '../../../../StyledButton';
-import ScreenLayout from '../../../common/components/ScreenLayout';
+import ScreenLayout from '../../components/ScreenLayout';
 import { getFiatOnRampAggNavbar } from '../../../../Navbar';
 import { strings } from '../../../../../../../locales/i18n';
 import { useTheme } from '../../../../../../util/theme';
-import { useRampSDK } from '../../../common/sdk';
-import ErrorViewWithReporting from '../../../common/components/ErrorViewWithReporting';
+import { useRampSDK } from '../../sdk';
+import ErrorViewWithReporting from '../../components/ErrorViewWithReporting';
 import Routes from '../../../../../../constants/navigation/Routes';
-import useAnalytics from '../../../common/hooks/useAnalytics';
-import useRampNetwork from '../../../common/hooks/useRampNetwork';
+import useAnalytics from '../../hooks/useAnalytics';
+import useRampNetwork from '../../hooks/useRampNetwork';
 import styles from './GetStarted.styles';
 import useRegions from '../../hooks/useRegions';
