+++ app/components/UI/Ramp/common/Views/GetStarted/index.ts
