+++ app/components/UI/Ramp/common/Views/Quotes/Quotes.constants.ts
