+++ app/components/UI/Ramp/common/Views/Quotes/Timer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { ActivityIndicator, View } from 'react-native';
 import { useStyles } from '../../../../../hooks/useStyles';
-import { useRampSDK } from '../../../common/sdk';
+import { useRampSDK } from '../../sdk';
 
 import Text from '../../../../../Base/Text';
 import styleSheet from './Quotes.styles';
