+++ app/components/UI/Ramp/common/hooks/useHandleSuccessfulOrder.ts
@@ -11,9 +11,9 @@ import NotificationManager from '../../../../../core/NotificationManager';
 import { addFiatOrder, FiatOrder } from '../../../../../reducers/fiatOrders';
 import { toLowerCaseEquals } from '../../../../../util/general';
 import useThunkDispatch from '../../../../hooks/useThunkDispatch';
-import { useRampSDK } from '../../common/sdk';
-import { getNotificationDetails, stateHasOrder } from '../../common/utils';
-import useAnalytics from '../../common/hooks/useAnalytics';
+import { useRampSDK } from '../sdk';
+import { getNotificationDetails, stateHasOrder } from '../utils';
+import useAnalytics from './useAnalytics';
 import { hexToBN } from '../../../../../util/number';
 import { selectAccounts } from '../../../../../selectors/accountTrackerController';
 import Routes from '../../../../../constants/navigation/Routes';
